requirements.txt
fastmri==0.2.0
fire==0.4.0
h5py==3.7.0
joblib==1.2.0
matplotlib==3.5.0
numpy==1.23.5
pandas==1.5.0
pytorch_lightning==1.7.6
PyYAML==5.4.1
PyYAML==6.0
scikit_learn==1.2.2
scikit_learn==1.0.2
torch==1.9.0+cu111
torchmetrics==0.9.3
tqdm==4.65.0
